[refactor/#5] 패키지 구조 변경
- core (designsystem, navigation, util) 폴더 설정
- model을 presentation 안으로 위치
