+++ app/src/main/java/org/sopt/at/core/designsystem/theme/Type.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.ui.theme
+package org.sopt.at.core.designsystem.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
