+++ app/src/main/java/org/sopt/at/core/navigation/MainTabRoute.kt
@@ -1,0 +1,3 @@
+package org.sopt.at.core.navigation
+
+interface MainTabRoute: Route
