+++ app/src/main/java/org/sopt/at/core/navigation/Route.kt
@@ -1,0 +1,3 @@
+package org.sopt.at.core.navigation
+
+interface Route
