+++ app/src/main/java/org/sopt/at/presentation/MainActivity.kt
@@ -5,7 +5,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import org.sopt.at.presentation.main.MainScreen
-import org.sopt.at.ui.theme.ATSOPTANDROIDTheme
+import org.sopt.at.core.designsystem.theme.ATSOPTANDROIDTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
