+++ app/src/main/java/org/sopt/at/presentation/auth/signup/SignUpMainScreen.kt
@@ -6,19 +6,17 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.lifecycle.viewmodel.compose.viewModel
 import org.sopt.at.R
 import org.sopt.at.presentation.auth.login.navigation.Login
-import org.sopt.at.ui.common.appbar.CommonTopAppBar
-import org.sopt.at.ui.theme.ATSOPTANDROIDTheme
+import org.sopt.at.core.designsystem.component.appbar.CommonTopAppBar
+import org.sopt.at.core.designsystem.theme.ATSOPTANDROIDTheme
 
 @Composable
 fun SignUpRoute(
