+++ app/src/main/java/org/sopt/at/presentation/auth/signup/SignUpPwdScreen.kt
@@ -25,11 +25,11 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.sopt.at.R
-import org.sopt.at.ui.common.button.ButtonSizeType
-import org.sopt.at.ui.common.button.CommonOutlinedButton
-import org.sopt.at.ui.common.textfield.CommonTextField
-import org.sopt.at.ui.common.textfield.TextFieldType
-import org.sopt.at.ui.theme.ATSOPTANDROIDTheme
+import org.sopt.at.core.designsystem.component.button.ButtonSizeType
+import org.sopt.at.core.designsystem.component.button.CommonOutlinedButton
+import org.sopt.at.core.designsystem.component.textfield.CommonTextField
+import org.sopt.at.core.designsystem.component.textfield.TextFieldType
+import org.sopt.at.core.designsystem.theme.ATSOPTANDROIDTheme
 
 @Composable
 fun SignUpPwdScreen(
