+++ app/src/main/java/org/sopt/at/presentation/home/component/BannerCarousel.kt
@@ -18,8 +18,8 @@ import androidx.compose.ui.unit.dp
 import com.skydoves.landscapist.ShimmerParams
 import com.skydoves.landscapist.glide.GlideImage
 import org.sopt.at.R
-import org.sopt.at.ui.theme.ButtonTint
-import org.sopt.at.ui.theme.TextFieldBg
+import org.sopt.at.core.designsystem.theme.ButtonTint
+import org.sopt.at.core.designsystem.theme.TextFieldBg
 
 @Composable
 fun BannerCarousel(
