+++ app/src/main/java/org/sopt/at/presentation/home/navigation/HomeNavigation.kt
@@ -6,8 +6,8 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import kotlinx.serialization.Serializable
+import org.sopt.at.core.navigation.MainTabRoute
 import org.sopt.at.presentation.home.HomeRoute
-import org.sopt.at.util.MainTabRoute
 
 fun NavController.navigateToHome(
     id: String,
