+++ app/src/main/java/org/sopt/at/presentation/search/SearchNavigation.kt
@@ -6,8 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import kotlinx.serialization.Serializable
-import org.sopt.at.presentation.history.HistoryRoute
-import org.sopt.at.util.MainTabRoute
+import org.sopt.at.core.navigation.MainTabRoute
 
 fun NavController.navigateToSearch(
     navOptions: NavOptions? = null
