+++ app/src/main/java/org/sopt/at/presentation/shorts/ShortsNavigation.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import kotlinx.serialization.Serializable
-import org.sopt.at.util.MainTabRoute
+import org.sopt.at.core.navigation.MainTabRoute
 
 fun NavController.navigateToShorts(
     navOptions: NavOptions? = null
