--- app/src/main/java/org/sopt/at/util/NavigationRoute.kt
@@ -1,5 +1,0 @@
-package org.sopt.at.util
-
-interface Route
-
-interface MainTabRoute: Route
